+++ test/majorver/test_dupes_linking.py
@@ -2,7 +2,7 @@ from . import *
 
 @mark.copyfiles(('fileA',2),('fileB',3))
 def test_link(copiedtestfiles):
-    dupes = DuplicateFiles.frompath(copiedtestfiles.root)
+    dupes = DuplicateFiles.frompaths(copiedtestfiles.root)
     dupes.link()
     fileAino = copiedtestfiles.paths['fileA'][0].stat().st_ino
     fileBino = copiedtestfiles.paths['fileB'][0].stat().st_ino
@@ -19,7 +19,7 @@ def test_link(copiedtestfiles):
 def test_link_duplicatefileswithmultiplegroupsoflinks(copiedtestfiles):
     """fileA and fileA-copy are identical and BOTH have multiple copies AND multiple hardlinks
     """
-    dupes = DuplicateFiles.frompath(copiedtestfiles.root)
+    dupes = DuplicateFiles.frompaths(copiedtestfiles.root)
     dupes.link()
     fileAino = copiedtestfiles.paths['fileA'][0].stat().st_ino
     inoscorrect = {(fileid, i): file.stat().st_ino == fileAino for fileid in ('fileA', 'fileA-copy') for i, file in enumerate(copiedtestfiles.paths[fileid])}
@@ -41,5 +41,5 @@ def test_donothingifonlylinks(copiedtestfiles, monkeypatch):
     monkeypatch.setattr(dupes, "_replacewithlink", _dontlink)
     # Monkeypatch was validated by adding an extra set of files which did need linking
     
-    duplicatefiles = DuplicateFiles.frompath(copiedtestfiles.root)
+    duplicatefiles = DuplicateFiles.frompaths(copiedtestfiles.root)
     duplicatefiles.link()